+++ anyblok/tests/test_column.py
@@ -3,6 +3,7 @@
 #    Copyright (C) 2014 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2019 Jean-Sebastien SUZANNE <js.suzanne@gmail.com>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -1164,6 +1165,18 @@ class TestColumns:
         assert test.col == 'john.smith@foo.com'
         assert registry.Test.query().filter_by(
             col='John.Smith@foo.com').count() == 1
+
+    @pytest.mark.skipif(not has_cryptography,
+                        reason="cryptography is not installed")
+    def test_email_with_encrypt_key(self):
+        registry = self.init_registry(simple_column, ColumnType=Email,
+                                      encrypt_key='secretkey')
+        test = registry.Test.insert(col='John.Smith@foo.com')
+        registry.session.commit()
+        assert test.col == 'john.smith@foo.com'
+        res = registry.execute('select col from test where id = %s' % test.id)
+        res = res.fetchall()[0][0]
+        assert res != test.col
 
     @pytest.mark.skipif(not has_pycountry, reason="pycountry is not installed")
     def test_pycoundtry_at_insert(self):