+++ anyblok/tests/test_column.py
@@ -1177,6 +1177,18 @@ class TestColumns:
         assert registry.Test.query().filter_by(
             col='John.Smith@foo.com').count() == 1
 
+    @pytest.mark.skipif(not has_cryptography,
+                        reason="cryptography is not installed")
+    def test_email_with_encrypt_key(self):
+        registry = self.init_registry(simple_column, ColumnType=Email,
+                                      encrypt_key='secretkey')
+        test = registry.Test.insert(col='John.Smith@foo.com')
+        registry.session.commit()
+        assert test.col == 'john.smith@foo.com'
+        res = registry.execute('select col from test where id = %s' % test.id)
+        res = res.fetchall()[0][0]
+        assert res != test.col
+
     @pytest.mark.skipif(not has_pycountry, reason="pycountry is not installed")
     def test_pycoundtry_at_insert(self):
         registry = self.init_registry(simple_column, ColumnType=Country)
@@ -1275,6 +1287,28 @@ class TestColumns:
         res = registry.execute('select col from test where id = %s' % test.id)
         res = res.fetchall()[0][0]
         assert res != test.col
+
+    def test_foreign_key_on_mapper_issue_112(self):
+
+        def add_in_registry():
+
+            @Declarations.register(Declarations.Model)
+            class Template:
+                code = String(primary_key=True, db_column_name='ProductId')
+
+            @Declarations.register(Declarations.Model)
+            class Item:
+                id = Integer(primary_key=True, db_column_name='ProductDetailId')
+                template_code = String(
+                    db_column_name='ProductId',
+                    foreign_key=Model.Template.use('code'))
+
+        registry = self.init_registry(add_in_registry)
+        registry.Template.insert(code='test')
+        registry.Item.insert(template_code='test')
+
+        with pytest.raises(Exception):
+            registry.Item.insert(template_code='other')
 
 
 class TestColumnsAutoDoc: