+++ doc/CHANGES.rst
@@ -19,6 +19,8 @@ CHANGELOG
 -----
 
 * Nom the Email field can be crypted
+* Removed compatibility with **Python 3.4**, because this version is
+  deprecated
 
 0.22.5 (2019-06-24)
 -------------------