+++ doc/CHANGES.rst
@@ -18,6 +18,7 @@ CHANGELOG
 1.0.0
 -----
 
+* Nom the Email field can be crypted
 * Removed compatibility with **Python 3.4**, because this version is
   deprecated