+++ doc/CHANGES.rst
@@ -18,7 +18,9 @@ CHANGELOG
 1.0.0
 -----
 
-* Nom the Email field can be crypted
+* Now the Email field can be crypted
+* Removed compatibility with **Python 3.4**, because this version is
+  deprecated
 
 0.22.5 (2019-06-24)
 -------------------