+++ doc/CHANGES.rst
@@ -15,6 +15,20 @@
 CHANGELOG
 =========
 
+1.0.0
+-----
+
+* Nom the Email field can be crypted
+* Removed compatibility with **Python 3.4**, because this version is
+  deprecated
+
+0.22.5 (2019-06-24)
+-------------------
+
+* Reverted console script **anyblok_nose**: to keep compatibility
+* Fixed minimum version of **SQLalchemy >= 1.3** in package dependencies
+* Removed some deprecation warning
+
 0.22.2 (2019-06-23)
 -------------------